Extract adjustTargetSource and create utils/
