+++ src/utils/SimpleAnnotationServerV2Adapter.ts
@@ -6,15 +6,15 @@
 
 */
 
-import type { Annotation as V2Annotation } from "./types/V2/Annotation";
-import type { Annotation as V3Annotation } from "./types/V3/Annotation";
-import type { AnnotationPage } from "./types/V3/AnnotationPage";
-import type { Body as V2Body } from "./types/V2/Body";
-import type { Body as V3Body } from "./types/V3/Body";
-import type { Selector as V2Selector } from "./types/V2/Selector";
-import type { Selector as V3Selector } from "./types/V3/Selector";
-import type { Source } from "./types/V3/Source";
-import type { Target } from "./types/V3/Target";
+import type { Annotation as V2Annotation } from "../types/V2/Annotation";
+import type { Annotation as V3Annotation } from "../types/V3/Annotation";
+import type { AnnotationPage } from "../types/V3/AnnotationPage";
+import type { Body as V2Body } from "../types/V2/Body";
+import type { Body as V3Body } from "../types/V3/Body";
+import type { Selector as V2Selector } from "../types/V2/Selector";
+import type { Selector as V3Selector } from "../types/V3/Selector";
+import type { Source } from "../types/V3/Source";
+import type { Target } from "../types/V3/Target";
 
 export default class SimpleAnnotationServerV2Adapter {
     canvasId: string;
