+++ tests/index.test.ts
@@ -1,5 +1,5 @@
 import AnnotationServerStorage from "../src/index";
-import SimpleAnnotationServerV2Adapter from "../src/SimpleAnnotationServerV2Adapter";
+import SimpleAnnotationServerV2Adapter from "../src/utils/SimpleAnnotationServerV2Adapter";
 
 // Mock the adapter
 jest.mock("../src/SimpleAnnotationServerV2Adapter", () => {
